refactor: extract status bar api to api package Signed-off-by: Florent Benoit <fbenoit@redhat.com>
