+++ packages/api/src/status-bar.ts
@@ -1,0 +1,36 @@
+/**********************************************************************
+ * Copyright (C) 2025 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ***********************************************************************/
+
+export const STATUS_BAR_UPDATED_EVENT_NAME = 'status-bar-updated';
+
+export interface StatusBarEntryDescriptor {
+  priority: number;
+  alignLeft: boolean;
+  entry: StatusBarEntry;
+}
+
+export interface StatusBarEntry {
+  text?: string;
+  tooltip?: string;
+  activeIconClass?: string;
+  inactiveIconClass?: string;
+  enabled: boolean;
+  command?: string;
+  commandArgs?: unknown[];
+  highlight?: boolean;
+}
