+++ packages/renderer/src/lib/statusbar/StatusBarItem.spec.ts
@@ -20,7 +20,8 @@ import { render, screen } from '@testing-library/svelte';
 import userEvent from '@testing-library/user-event';
 import { beforeAll, beforeEach, expect, test, vi } from 'vitest';
 
-import type { StatusBarEntry } from '../../../../main/src/plugin/statusbar/statusbar-registry';
+import type { StatusBarEntry } from '/@api/status-bar';
+
 import { iconClass } from './StatusBarItem';
 import StatusBarItem from './StatusBarItem.svelte';
